Make WMSHistory and MonitoringHistory corrector classes derived from BaseHistory parent class
